trip-planner/requirements.txt
streamlit>=1.20
python-dotenv>=1.0
requests>=2.28
openai>=1.0
reportlab>=4.0
pydantic>=1.10
tenacity>=8.0
pytest>=7.0
langgraph>=0.1
